// apps/stm/src/features/admin/courseApprovalDecisions/interfaces.ts
export interface CourseApprovalDecision {
  maHocVien?: string;
  hoTen?: string;
  gioiTinh?: string;
  ngaySinh?: string;
  soDienThoai?: string;
  email?: string;
  maKhoaHoc?: string;
  tenKhoaHoc?: string;
  cumThoiGian?: string;
  chiNhanh?: string;
  loaiQuyetDinh?: string;
  ngayQuyetDinh?: string;
  tenQuyetDinh?: string;
  ghiChu?: string;
  minhChungQuyetDinh?: string;
  bienLaiChuyenKhoan?: string;
  huyQuyetDinh?: boolean;
}